+++ src/main/java/roomescape/persistence/repository/ThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.persistence;
+package roomescape.persistence.repository;
 
 import java.util.List;
 import roomescape.model.Theme;
